Refactoring to move release date getters into their own module
