+++ server/routes/approvingLicences/handlers/approvalCases.ts
@@ -3,13 +3,9 @@ import _ from 'lodash'
 import { getUnixTime } from 'date-fns'
 import CaseloadService from '../../../services/caseloadService'
 import PrisonerService from '../../../services/prisonerService'
-import {
-  convertToTitleCase,
-  releaseDateLabel,
-  selectReleaseDate,
-  selectReleaseDateFromLicence,
-} from '../../../utils/utils'
+import { convertToTitleCase, releaseDateLabel } from '../../../utils/utils'
 import LicenceService from '../../../services/licenceService'
+import { getReleaseDateFromLicence, getReleaseDateFromNomis } from '../../../services/getReleaseDate'
 
 export default class ApprovalCaseRoutes {
   constructor(
@@ -43,7 +39,7 @@ export default class ApprovalCaseRoutes {
           name: convertToTitleCase(`${c.nomisRecord.firstName} ${c.nomisRecord.lastName}`.trim()),
           prisonerNumber: c.nomisRecord.prisonerNumber,
           probationPractitioner: c.probationPractitioner,
-          releaseDate: licence ? selectReleaseDateFromLicence(licence) : selectReleaseDate(c.nomisRecord),
+          releaseDate: licence ? getReleaseDateFromLicence(licence) : getReleaseDateFromNomis(c.nomisRecord),
           releaseDateLabel: releaseDateLabel(licence, c.nomisRecord),
         }
       })
