+++ server/utils/utils.ts
@@ -1,6 +1,6 @@
 import moment, { Moment } from 'moment'
 import { Holiday } from 'uk-bank-holidays'
-import { format, isBefore, parse } from 'date-fns'
+import { isBefore, parse } from 'date-fns'
 import AuthRole from '../enumeration/authRole'
 import SimpleDateTime from '../routes/creatingLicences/types/simpleDateTime'
 import SimpleDate from '../routes/creatingLicences/types/date'
@@ -8,7 +8,6 @@ import SimpleTime, { AmPm } from '../routes/creatingLicences/types/time'
 import Address from '../routes/creatingLicences/types/address'
 import { Licence, LicenceSummary } from '../@types/licenceApiClientTypes'
 import { Prisoner } from '../@types/prisonerSearchApiClientTypes'
-import logger from '../../logger'
 import { PrisonApiPrisoner } from '../@types/prisonApiClientTypes'
 
 const properCase = (word: string): string =>
@@ -171,51 +170,6 @@ const isBankHolidayOrWeekend = (date: Moment, bankHolidays: Holiday[]) => {
 const licenceIsTwoDaysToRelease = (licence: Licence) =>
   moment(licence.conditionalReleaseDate, 'DD/MM/YYYY').diff(moment(), 'days') <= 2
 
-const selectReleaseDate = (nomisRecord: Prisoner) => {
-  let dateString = nomisRecord.conditionalReleaseDate
-
-  if (nomisRecord.confirmedReleaseDate) {
-    dateString = nomisRecord.confirmedReleaseDate
-  }
-
-  if (nomisRecord.conditionalReleaseOverrideDate) {
-    dateString = nomisRecord.conditionalReleaseOverrideDate
-  }
-
-  if (!dateString) {
-    return 'not found'
-  }
-
-  try {
-    dateString = format(new Date(dateString), 'dd MMM yyyy')
-  } catch (e) {
-    logger.error(
-      `Invalid date error: ${e.message} for prisonerNumber: ${nomisRecord.prisonerNumber} using date: ${dateString}`
-    )
-  }
-
-  return dateString
-}
-
-const selectReleaseDateFromLicence = (licence: Licence) => {
-  let dateString = licence.actualReleaseDate || licence.conditionalReleaseDate
-
-  if (!dateString) {
-    logger.error(`No release date found for NOMIS ID: ${licence.nomsId}`)
-    return 'not found'
-  }
-
-  dateString = toDateString(dateString)
-
-  try {
-    dateString = format(new Date(dateString), 'dd MMM yyyy')
-  } catch (e) {
-    logger.error(`Invalid date error: ${e.message} for NOMIS ID: ${licence.nomsId} using date: ${dateString}`)
-  }
-
-  return dateString
-}
-
 const isPassedArdOrCrd = (licence: LicenceSummary, prisoner: Prisoner | PrisonApiPrisoner): boolean => {
   const releaseDate =
     prisoner.legalStatus !== 'IMMIGRATION_DETAINEE'
@@ -258,8 +212,6 @@ export {
   formatAddress,
   isBankHolidayOrWeekend,
   licenceIsTwoDaysToRelease,
-  selectReleaseDate,
-  selectReleaseDateFromLicence,
   isPassedArdOrCrd,
   releaseDateLabel,
   toDateString,
